requirements.txt
isort
black
flake8
mypy
pyyaml
pydantic
python-dotenv
gradio
torch
transformers
accelerate